+++ bittrade_binance_websocket/framework/framework.py
@@ -53,10 +53,9 @@ from bittrade_binance_websocket.rest.listen_key import (
     delete_listen_key_http_factory,
     get_active_listen_key_http_factory,
     get_listen_key_http_factory,
-    isolated_margin_delete_listen_key_http_factory,
-    isolated_margin_get_active_listen_key_http_factory,
-    isolated_margin_get_listen_key_http_factory,
-    isolated_margin_ping_listen_key_http_factory,
+    margin_delete_listen_key_http_factory,
+    margin_get_listen_key_http_factory,
+    margin_ping_listen_key_http_factory,
     ping_listen_key_http_factory,
 )
 from bittrade_binance_websocket.rest.margin_loan import (
@@ -99,14 +98,14 @@ def get_framework(
     get_listen_key_http = get_listen_key_http_factory(user_stream_signer_http)
     keep_alive_listen_key_http = ping_listen_key_http_factory(user_stream_signer_http)
     delete_listen_key_http = delete_listen_key_http_factory(user_stream_signer_http)
-    isolated_margin_get_listen_key_http = (
-        isolated_margin_get_active_listen_key_http_factory(user_stream_signer_http)
+    margin_get_listen_key_http = (
+        margin_get_listen_key_http_factory(user_stream_signer_http)
     )
-    isolated_margin_keep_alive_listen_key_http = (
-        isolated_margin_ping_listen_key_http_factory(user_stream_signer_http)
+    margin_keep_alive_listen_key_http = (
+        margin_ping_listen_key_http_factory(user_stream_signer_http)
     )
-    isolated_margin_delete_listen_key_http = (
-        isolated_margin_delete_listen_key_http_factory(user_stream_signer_http)
+    margin_delete_listen_key_http = (
+        margin_delete_listen_key_http_factory(user_stream_signer_http)
     )
 
     # Setup up public stream
@@ -131,9 +130,13 @@ def get_framework(
         connection_operators.keep_messages_only()
     )
 
-    def isolated_margin_user_stream_factory(symbol: str):
-        key_getter = lambda: isolated_margin_get_listen_key_http(symbol)
-        keep_alive = lambda key: isolated_margin_keep_alive_listen_key_http(key, symbol)
+    def margin_user_stream_factory(symbol: str | None=None):
+        """
+        Skip symbol for cross margin
+        """
+        symbol = symbol or ""
+        key_getter = lambda: margin_get_listen_key_http(symbol)
+        keep_alive = lambda key: margin_keep_alive_listen_key_http(key, symbol)
         socket_bundles = private_websocket_user_stream(key_getter, keep_alive)
         socket = socket_bundles.pipe(connection_operators.keep_new_socket_only())
 
@@ -177,8 +180,8 @@ def get_framework(
         get_account_information_http=get_account_information_http,
         get_active_listen_key_http=get_active_listen_key_http,
         get_listen_key_http=get_listen_key_http,
-        isolated_margin_get_listen_key_http=isolated_margin_get_listen_key_http,
-        isolated_margin_user_stream_factory=isolated_margin_user_stream_factory,
+        margin_get_listen_key_http=margin_get_listen_key_http,
+        margin_user_stream_factory=margin_user_stream_factory,
         keep_alive_listen_key_http=keep_alive_listen_key_http,
         market_symbol_price_ticker_http=symbol_price_ticker_http,
         market_symbol_price_book_ticker_http=symbol_price_book_ticker_http,