+++ bittrade_binance_websocket/rest/account_information.py
@@ -11,15 +11,18 @@ from bittrade_binance_websocket.rest.http_factory_decorator import http_factory
 
 @http_factory(dict)
 def get_account_information_http_factory(
-    isolated: bool = False, symbols: list[str] | None = None
+    is_margin: bool=False, is_isolated: bool = False, symbols: list[str] | None = None
 ):
     params = {}
-    if isolated and symbols:
+    if is_isolated and symbols:
         params["symbols"] = ",".join(symbols)
+    endpoint = endpoints.BinanceEndpoints.ACCOUNT_INFORMATION
+    if is_margin:
+        endpoint = endpoints.BinanceEndpoints.QUERY_CROSS_MARGIN_ACCOUNT_DETAILS
+        if is_isolated:
+            endpoint = endpoints.BinanceEndpoints.QUERY_ISOLATED_MARGIN_ACCOUNT_DETAILS
     return request.RequestMessage(
         method="GET",
-        endpoint=endpoints.BinanceEndpoints.MARGIN_ACCOUNT_INFORMATION
-        if isolated
-        else endpoints.BinanceEndpoints.ACCOUNT_INFORMATION,
+        endpoint=endpoint,
         params=params,
     )